Crawl-URL/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==2.0.4
chromedriver-autoinstaller==0.6.4
idna==3.2
packaging==23.2
PyYAML==6.0.1
requests==2.31.0
selenium==3.141.0
soupsieve==2.2.1
urllib3==1.26.6
wget==3.2
pandas